echo

+++ Builtins/unset.c
@@ -6,7 +6,7 @@
 /*   By: moudrib <moudrib@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/22 12:48:36 by moudrib           #+#    #+#             */
-/*   Updated: 2023/05/24 15:04:57 by moudrib          ###   ########.fr       */
+/*   Updated: 2023/06/13 10:39:01 by moudrib          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,11 +78,19 @@ void	remove_variable(t_vars *v, t_env **env)
 {
 	v->i = 0;
 	v->temp1 = *env;
+	if (ft_lstsize_env(*env) == 1)
+	{
+		ft_destroy_list_env(env);
+		return ;
+	}
 	while (v->temp1)
 	{
 		v->i++;
 		if (v->temp1 && !ft_strcmp(v->var, v->temp1->key))
+		{
 			delete_node(env, v->i);
+			return ;
+		}
 		v->temp1 = v->temp1->link;
 	}
 }
@@ -91,14 +99,10 @@ void	unset(t_list **list, t_env **env)
 {
 	t_vars	v;
 
-	v.tmp1 = *list;
 	v.vars = 0;
-	while (v.tmp1)
-	{
-		if (!ft_strcmp(v.tmp1->type, "PIPE"))
-			return ;
-		v.tmp1 = v.tmp1->link;
-	}
+	v.tmp1 = *list;
+	if (search_for_pipe(&v))
+		return ;
 	if (!env || !(*env) || ft_lstsize_env(*env) == 0)
 		return ;
 	v.tmp1 = lexer_for_unset(list);
